Ahora llama al paquete de consultas y usa una función

+++ demoproyecto1BDII/src/main/java/com/oracle/demoproyecto1BDII/model/VentaClienteDTO.java
@@ -1,23 +1,25 @@
 package com.oracle.demoproyecto1BDII.model;
 
+import java.math.BigDecimal;
+
 public class VentaClienteDTO {
-    private int idCliente;
+    private long idCliente;
     private String cliente;
-    private int totalVendido;
+    private BigDecimal totalVendido;
     private int itemsComprados;
 
-    public VentaClienteDTO(int idCliente, String cliente,int itemsComprados, int totalVendido) {
+    public VentaClienteDTO(long idCliente, String cliente,int itemsComprados, BigDecimal totalVendido) {
         this.idCliente = idCliente;
         this.cliente = cliente;
         this.totalVendido = totalVendido;
         this.itemsComprados = itemsComprados;
     }
 
-    public int getIdCliente() {
+    public long getIdCliente() {
         return idCliente;
     }
 
-    public void setIdCliente(int idCliente) {
+    public void setIdCliente(long idCliente) {
         this.idCliente = idCliente;
     }
 
@@ -29,11 +31,11 @@ public class VentaClienteDTO {
         this.cliente = cliente;
     }
 
-    public int getTotalVendido() {
+    public BigDecimal getTotalVendido() {
         return totalVendido;
     }
 
-    public void setTotalVendido(int totalVendido) {
+    public void setTotalVendido(BigDecimal totalVendido) {
         this.totalVendido = totalVendido;
     }
